requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
redis==4.6.0
openai==0.28.1
langchain==0.0.335
langchain-openai==0.0.2
faiss-cpu==1.7.4
PyPDF2==3.0.1
python-dotenv==1.0.0
sentence-transformers==2.2.2
numpy==1.24.3
tiktoken==0.5.1
requests==2.31.0